s4c30: just implement md4 and length extension still works
MD4 is using little endian, and almost everything else is big endian. So
there are a lot of different problem points, but the most important one
was: converting the input using my self written byte to u32 converter.

Had to compare the source code of mine and another rust implementation 3
times before I decided to reimplement that and suddenly the tests were
green.