c31: Started implementing timing sidechannel attack
Sadly the detection is not stable. somewhere around 10-15 characters if
fails and can't recover. I used only 15-25ms slowdown, since I did not
feel like waiting forever (>15 min). But that also means the detection
is more susceptible to other influences. And I'm not using the specific
timing increase in my detection for early returns. So there are some
more things left to try.