Refactor bruteforce and use rayon to shorten runtime
The majority of time is taken up by the intentional sleep instructions
in the comparison. So it is easy to have even more threads than cores
and just do all possible values of the next byte at once. This reduces
the runtime from more than 40 minutes to less than 15 seconds. And using
rayon means the only change necessary is to go from `.into_iter()` to
`.into_par_iter()`.